+++ oonipipeline/src/oonipipeline/models/observations.py
@@ -1,8 +1,7 @@
-from dataclasses import dataclass, field
 import dataclasses
+from dataclasses import dataclass, field
 from datetime import datetime
 from typing import (
-    NamedTuple,
     Optional,
     List,
     Tuple,
